Test for overview stats title util function.
