+++ app/assets/javascripts/utils/article_utils.js
@@ -58,19 +58,4 @@ export default class ArticleUtils {
       wikiversity: 'collection'
     }
   };
-
-  // Get tabs and stats title for tabbed course overview stats.
-  // Here, wiki_ns_key is key of a course_stats object, which 
-  // identifies wiki or wiki-namespace of stats,
-  // eg.: 'en.wikibooks.org-namespace-102', 'www.wikidata.org'
-  static overviewStatsTitle(wiki_ns_key) {
-    // If stats is for wikidata overview, directly return the wiki domain
-    if (!wiki_ns_key.includes('namespace')) return wiki_ns_key;
-    const project = wiki_ns_key.split('.')[1];
-    const wiki_domain = wiki_ns_key.split('-')[0];
-    const ns_id = wiki_ns_key.split('-')[2];
-    let ns_title = ArticleUtils.NamespaceIdMapping[ns_id];
-    if (typeof (ns_title) !== 'string') ns_title = ns_title[project];
-    return `${wiki_domain} - ${I18n.t(`namespace.${ns_title}`)}`;
-  };
 }
